Moved agent to ins own folder
