+++ python-code/src/amr_hub_abm/agent/__init__.py
@@ -1,0 +1,1 @@
+"""Agent module for the AMR Hub ABM simulation."""
