+++ python-code/src/amr_hub_abm/simulation.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 
     from matplotlib.axes import Axes
 
-    from amr_hub_abm.agent import Agent
+    from amr_hub_abm.agent.agent import Agent
     from amr_hub_abm.space.building import Building
     from amr_hub_abm.space.room import Room
