+++ python-code/src/amr_hub_abm/space/building.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from matplotlib.axes import Axes
 
-    from amr_hub_abm.agent import Agent
+    from amr_hub_abm.agent.agent import Agent
     from amr_hub_abm.space.floor import Floor
