+++ python-code/src/amr_hub_abm/space/content.py
@@ -17,7 +17,7 @@ from typing import TYPE_CHECKING
 import shapely
 
 if TYPE_CHECKING:
-    from amr_hub_abm.agent import AgentType
+    from amr_hub_abm.agent.agent import AgentType
     from amr_hub_abm.space.location import Location
